Optimization of  import-recompute-logic.php. 

We've adopted phpIPAM for a fairly large internal network (~7500 subnets added so far and growing). The biggest wins so far in addressing the slow performance feedback from users has been enabling MySQL query caching and migrating to php7 from php5.5.  It may be worth advocating php7 on the homepage for large installations as the improvements are significant.

Subnets::fetch_subnet_slaves_recursive is the worst performing function remaining taking 45 seconds to render the 10.0.0.0/8 subnets details page (94.9% of total time).  I'm leaning towards a MySQL stored procedure to generate a temporary table containing a list of child subnet ids for a given subnet. We can then avoid the slow recursion in php code ( Subnets::fetch_subnet_slaves called 21,143 times. )

    # Subnet 10.0.0.0/8 has id 1000
    CALL SubnetFamilyTreeById(1000);
    # The temporary table subnet_family_tree1000 now contains a list of 10.0.0.0/8 children ids.
    SELECT * FROM subnets WHERE id IN (SELECT id FROM subnet_family_tree1000);

Is this something that you or others are already working on and would you be happy with such an approach?

The nested set model has the potential to be even faster but is much harder to implement, has issues with insertion and deletion performance and would break if the database was modified outside phpIPAM.

p.s. I'm planning on revisiting import-recompute-logic.php later to extend the form so the users can pick from the available candidates master subnets from a drop-down box.

Thanks

-----------------------------------------------
Improve the performance of import-recompute-logic.php. The majority of the speed
up is from avoiding searching subnets of equal or larger mask size when looking
for a new master subnet e.g. When looking for the master of 10.100.0.0/16 there
is no point looking at /24 subnets, we only need to search /15, /14, /13, /12...

When selecting a new master subnet from available candidates give preference to
subnets in the same VRF, followed by the currently set master subnet.

The largest observed speed increase was from the migration from php55 to php70.
This also has the added benefit of reducing memory_get_peak_usage() by ~%50.

Benchmarks below are from a live production database (~7500 subnets). Synthetic
datasets show greater improvements.

RedHat 6.7. Xeon @2.60GHz, MySQL 5.1.73

PHP 5.5.21, execute import-recompute-logic.php
Original:   112 seconds, 95.7Mb Peak memory usage
Patched :    62 seconds, 77.5Mb Peak memory usage

PHP 7.0.10, execute import-recompute-logic.php
Original:    46 seconds, 36.6Mb Peak memory usage
Patched:      8 seconds, 37.4Mb Peak memory usage