@malaiam 2017-05-04
Hey!

I'm glad someone took the time to look into the optimization of that part. I wonder what is the use case which requires the execution of the master/slave recomputation more than once, at the migration time. The intended use was to execute the recomputation only after the first import, while the rest of the adding / subnetting operations, within phpipam, would add the subnets in the correct master and the execution of the function would not be needed.

That being said, any improvement is welcomed. So I read your proposal and the bucket logic seems sound. I looked closely through the PHP code and that appears to implement the proposed improvement. Now, I didn't run any tests to look for bugs or corner cases, as I don't have any environment prepared - I'm not using phpipam actively these days, but I trust you did that extensively, right?

As for the further improvements via the MySQL procedure and temporary table, I cannot comment on it as I am not sure how much will that change the DB structure and/or load. Again, the first question to answer is if there are use cases to motivate such a change.

Regards,
Marius