@GaryAllan 2017-05-05
Hi

The use-case for the performance patches is usability and scalability. We have 3000+ sites with IP address information largely held in spreadsheets and we would like the information to be centralised in phpIPAM. The feature set of phpIPAM closely matched our requirements of multi-section, vrf aware and importing from spreadsheets.

We're slowly documenting everything manually via the spreadsheet import functionality and then running import-recompute-logic.php to correctly nest the results. We currently have around 7500 subnets documented and running recompute-logic is taking 2.5+minutes, with the patched recompute-logic and migration to php7 it now executes in <10s. This really helps the guys doing the batch importing.

We've been running the optimised patches in production with no issues and I've also tested functionality on a synthetically generated database containing 8000 IPv4 and 8000 IPv6 nested subnets. I've also switched between the original and optimised version of the function on the development instance and the ordering remained unchanged (They both generate identical nested trees).

I've submitted a separate pull request for the temporary table performance patches. I'll add more details to that thread.

Thanks