@RO-29 2017-05-08
Hi @GaryAllan 

you mentioned that 

```Subnets::fetch_subnet_slaves_recursive is the worst performing function remaining taking 45 seconds to render the 10.0.0.0/8 subnets details page (94.9% of total time). I'm leaning towards a MySQL stored procedure to generate a temporary table containing a list of child subnet ids for a given subnet. We can then avoid the slow recursion in php code ( Subnets::fetch_subnet_slaves called 21,143 times. )```

I am also facing a similar issue as have huge network (~40000 subnets) and this kills a lot of time.

Were you able to address this? Can you share your approach?

Thanks