@GaryAllan 2017-05-08
Hello,

I've submitted performance patches for review https://github.com/phpipam/phpipam/pull/1098 .  Are you able to test the pull request and provide feedback? I have further patches lined up based on the same approach. Feedback from users with large datasets would be useful.

Enabling the MySQL query cache also helps..... "query_cache_limit" is the maximum query result size that will be cached. "query_cache_size" is the total cache size. Given the size of your dataset you may need to increase these values but there are warnings online that increasing them too much can actually hurt performance and cause other issues such as query stalls when the large cache is flushed.  You can check the cache status with  SHOW STATUS LIKE 'qcache%';

    [mysqld]
    query_cache_type=1
    query_cache_size=64M
    query_cache_limit=8M

Thanks