@RO-29 2017-05-08
Yet to test this, thanks for sharing your approach.

Will try to test this tomorrow morning and let you know !

What optimal cache size do you suggest for my kind of dataset? (~40K records)