@GaryAllan 2017-05-08
Hi,

Impossible to say.... Start with the above settings and run SHOW STATUS LIKE 'qcache%'; Watch the Qcache_free_memory and Qcache_lowmem_prunes values.  Query caching in MariaDB 10.1 has been improved since the MySQL 5.1 days and on all versions performance decreases as you increase MySQL CPUs due to mutex locking so you''ll need to test on your tin. If you have a monster 32 core MySQL server it may actually make things worse :-)

I'd also highly recommend upgrading to PHP7 if you're on a PHP5.x release. It is significantly faster due to array() improvements and uses about 30-50% of the memory for the phpIPAM codebase.

    MariaDB [phpipam]> show status like 'qcache%';
    +-------------------------+----------+
    | Variable_name           | Value    |
    +-------------------------+----------+
    | Qcache_free_blocks      | 1        |
    | Qcache_free_memory      | 46153888 |
    | Qcache_hits             | 170395   |
    | Qcache_inserts          | 26824    |
    | Qcache_lowmem_prunes    | 0        |
    | Qcache_not_cached       | 346      |
    | Qcache_queries_in_cache | 10697    |
    | Qcache_total_blocks     | 21410    |
    +-------------------------+----------+
8 rows in set (0.01 sec)